+++ src/components/SlotsGames.vue
@@ -15,7 +15,7 @@
       </div>
       
       <div class="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-5 gap-4">
-        <GameMediaCard
+        <GameBannerCard
           v-for="slot in slotsGames"
           :key="slot.id"
           :title="slot.title"
@@ -41,7 +41,7 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import GameMediaCard from './GameMediaCard.vue'
+import GameBannerCard from './GameBannerCard.vue'
 import SlotsGamesModal from '../views/site/SlotsGamesModal.vue'
 
 const { t } = useI18n()
